# Transition to new CSRMatrix Interface