requires:
https://github.com/JuliaOpt/LinQuadOptInterface.jl/pull/101